requirements.txt
openai
boto3
python-dotenv
pandas
numpy
cryptography
requests>=2.32          # keep everything else happy
websocket-client>=1.8
six>=1.17

# Coinbase live data
cbpro2                  # drop the obsolete cbpro
